+++ packages/google_maps_places/google_maps_places/lib/google_maps_places.dart
@@ -2,35 +2,16 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-// ignore_for_file: avoid_classes_with_only_static_members
+library google_maps_places;
 
 import 'package:google_maps_places_platform_interface/google_maps_places_platform_interface.dart';
 
 export 'package:google_maps_places_platform_interface/google_maps_places_platform_interface.dart'
-    show AutocompletePrediction, LatLng, LatLngBounds, TypeFilter, PlaceType;
+    show
+        AutocompletePrediction,
+        LatLng,
+        LatLngBounds,
+        TypeFilter,
+        PlaceType;
 
-/// Document this
-class GoogleMapsPlaces {
-  static final GoogleMapsPlacesPlatform _instance =
-      GoogleMapsPlacesPlatform.instance;
-
-  /// Comment
-  static Future<List<AutocompletePrediction>> findAutocompletePredictions({
-    required String query,
-    LatLngBounds? locationBias,
-    LatLngBounds? locationRestriction,
-    LatLng? origin,
-    List<String?>? countries,
-    List<int?>? typeFilter,
-    bool? refreshToken,
-  }) async {
-    return await _instance.findAutocompletePredictions(
-        query: query,
-        countries: countries,
-        origin: origin,
-        locationBias: locationBias,
-        locationRestriction: locationRestriction,
-        typeFilter: typeFilter,
-        refreshToken: refreshToken);
-  }
-}
+part 'src/google_maps_places.dart';
