+++ packages/google_maps_places/google_maps_places_ios/lib/src/google_maps_places_ios.dart
@@ -1,0 +1,41 @@
+// Copyright 2013 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+part of google_maps_places_ios;
+
+/// An implementation of [GoogleMapsPlacesPlatform] for IOS.
+class GoogleMapsPlacesIOS extends GoogleMapsPlacesPlatform {
+  final GoogleMapsPlacesApiIOS _api = GoogleMapsPlacesApiIOS();
+
+  /// Registers this class as the default platform implementation.
+  static void registerWith() {
+    GoogleMapsPlacesPlatform.instance = GoogleMapsPlacesIOS();
+  }
+
+  @override
+  Future<List<AutocompletePrediction>> findAutocompletePredictions({
+    required String query,
+    LatLngBounds? locationBias,
+    LatLngBounds? locationRestriction,
+    LatLng? origin,
+    List<String?>? countries,
+    List<int?>? typeFilter,
+    bool? refreshToken,
+  }) async {
+    final List<AutocompletePredictionIOS?>? response =
+        await _api.findAutocompletePredictionsIOS(
+            query,
+            convertLatLngBounds(locationBias),
+            convertLatLngBounds(locationBias),
+            convertLatLng(origin),
+            countries,
+            typeFilter,
+            refreshToken);
+    if (response == null) {
+      throw ArgumentError(
+          'API returned empty response. Check log for details.');
+    }
+    return convertReponse(response);
+  }
+}
