+++ packages/google_maps_places/google_maps_places_platform_interface/lib/src/types/place_type.dart
