+++ packages/google_maps_places/google_maps_places_platform_interface/test/google_maps_places_platform_interface_test.dart
@@ -4,7 +4,7 @@
 
 import 'package:flutter_test/flutter_test.dart';
 import 'package:google_maps_places_platform_interface/google_maps_places_platform_interface.dart';
-import 'package:google_maps_places_platform_interface/method_channel/method_channel_google_maps_places.dart';
+import 'package:google_maps_places_platform_interface/src/method_channel/method_channel_google_maps_places.dart';
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
